+++ grelmicro/trace/errors.py
@@ -1,7 +1,11 @@
 """Tracing Errors."""
 
-from grelmicro.errors import GrelmicroError
+from grelmicro.errors import GrelmicroError, SettingsValidationError
 
 
 class TracingError(GrelmicroError):
     """Base tracing error."""
+
+
+class TracingSettingsValidationError(TracingError, SettingsValidationError):
+    """Tracing Settings Validation Error."""